+++ Doberman/__init__.py
@@ -8,6 +8,9 @@ from .BaseMonitor import *
 from .BaseSensor import *
 from .Database import *
 from .AlarmMonitor import *
+from .Node import *
+from .AlarmNodes import *
+from .ControlNodes import *
 from .HostMonitor import *
 from .SensorMonitor import *
 from .PipelineMonitor import *
